+++ ui/litellm-dashboard/src/app/globals.css
@@ -33,12 +33,13 @@ body {
   }
 
   .indigo-button-3d {
-    background: linear-gradient(#7a75ff, #4f46f1);
+    /* background: linear-gradient(#7a75ff, #4f46f1);
     box-shadow:
       inset 0 -1px 1px rgba(0, 0, 0, 0.08),
       inset 0 1px 1px rgba(255, 255, 255, 0.4),
       0 0 0px 1px #4d45e8,
-      0 4px 3px -1px rgba(0, 0, 0, 0.15);
+      0 4px 3px -1px rgba(0, 0, 0, 0.15); */
+    @apply bg-indigo-500;
   }
 }
