+++ ui/litellm-dashboard/src/components/virtual_keys_page/delete-virtual-key-dialog.tsx
